% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelMap.R
\name{skelMap}
\alias{skelMap}
\title{Plot colour-graded skeletal diagram.}
\usage{
skelMap(data, plot.field = "NISP", join.field = "El_code",
  template = "Bos", background = T, details = T, bg = "gray80",
  null.col = "white")
}
\arguments{
\item{data}{Data frame with at least one numeric or integer column containing data to be plotted,
and at least one field corresponding to the element names or codes used in the shapefile attribute table.}

\item{plot.field}{Character. The name of the field in 'data' containing the data to be plotted. Defaults to "NISP"}

\item{join.field}{Character. The name of the field in 'data' that should be used to link to the shapefile attribute table. Defaults to "El_code"}

\item{template}{Character. The name of the skeletal template to be used. This should be an .Rda file in the package "data" directory. Built-in
options are "Bos", "Ovis", "Equus", "Cervus", "Canis", "Sus". Defauls to "Bos".}

\item{background}{Logical: should a silhouette of the animal be plotted in the background? Defaults to TRUE.}

\item{details}{Logical: should details of bone features be drawn, as well as the element outlines? Defaults to TRUE.}

\item{bg}{Character. Colour for background silhouette, if 'background' is TRUE. Defaults to "gray80".}

\item{null.col.}{Colour for elements with missing data. Defaults to "white".}
}
\value{
None.
}
\description{
A wrapper for raster::spplot. Plots single column from input data frame on one of the built-in skeletal shapefiles.
}
\examples{
egData <- data.frame(El_code = c(1:140), NISP = rep(1:2, 70))
skelMap(egData)
}

